requirements.txt
fastapi
uvicorn
pydantic
python-multipart
jinja2
beautifulsoup4
pdfminer.six
sqlmodel
